+++ frontend/package.json
@@ -25,6 +25,7 @@
     "react-helmet": "^6.1.0",
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.28.0",
+    "react-toastify": "^11.0.5",
     "tailwind-merge": "^2.5.2",
     "tailwind-scrollbar-hide": "^2.0.0",
     "tailwindcss-animate": "^1.0.7"
